finishing up

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+
+    // to not let guest see edit, create pages
+    public function __construct()
+    {
+        // non loged in users will not be able to see posts pages except index and show
+        $this->middleware('auth', [ 'except' => ['index', 'show']]);
+        // $this->middleware('auth')->except(['index', 'show']); newer version
+    }
     /**
      * Display a listing of the resource.
      *
@@ -43,14 +54,33 @@ class PostsController extends Controller
         //validate and store the blog post
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload 
+        if($request->hasfile('cover_image')) {
+            //get file name with extention
+            // $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            // $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $fileName = $request->file('cover_image')->getClientOriginalName();
+            //get just extention
+            $extention = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $fileName. '_' .time().'.'.$extention;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        };
 
         // Create Post
 
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->cover_image = $fileNameToStore;
         $post->user_id = auth()->user()->id;
         $post->save();
 
@@ -79,6 +109,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // check for the correct user
+        if ( Auth::id() != $post->user_id ) {
+            return redirect('/posts')->with('error', 'Unauhorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -97,11 +131,28 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+        if($request->hasfile('cover_image')) {
+            //get file name with extention
+            // $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            // $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $fileName = $request->file('cover_image')->getClientOriginalName();
+            //get just extention
+            $extention = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $fileName. '_' .time().'.'.$extention;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         // Update Post
         $post = Post::find($id); // this is eloquent code
         $post->title = $request->input('title');
         $post->body = $request->input('body');
-
+        if($request->hasfile('cover_image')) {
+            Storage::delete('public/cover_images/'.$post->cover_image);
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -116,6 +167,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // check for the correct user
+        if ( Auth::id() != $post->user_id ) {
+            return redirect('/posts')->with('error', 'Unauhorized page');
+        }
+
+        //deleting the post image
+        if ($post->cover_image != 'noimage.jpg') {
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Delete');
